+++ packages/core/mashroom/src/config/mashroom_default_config.ts
@@ -18,6 +18,7 @@ const config: MashroomServerConfig = {
     }],
     ignorePlugins: [],
     indexPage: '/',
+    devModePreferredBuildTool: 'nx',
     devModeNpmExecutionTimeoutSec: undefined,
 };